Manual Scoring: Save and jump to next student Scoring: permit half points for all question types
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@53273 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/TestQuestionPool/classes/class.assTextQuestion.php
@@ -1018,7 +1018,7 @@ class assTextQuestion extends assQuestion implements ilObjQuestionScoringAdjusta
 				$affectedRows = $ilDB->manipulateF(
 					"INSERT INTO qpl_a_essay (answer_id, question_fi, answertext, points) 
 					 VALUES (%s, %s, %s, %s)",
-					 array('integer','integer','text','integer'),
+					 array('integer','integer','text','float'),
 					 array($next_id, $this->getId(), $row["answertext"], $row["points"])
 				);
 			}
